Storageclass not correct for a9s PAAS
I discovered that the storageclass in use for tutorial demonstrated on
the a9s PAAS would not successfully deploy all 3 PostgreSQL replicas of
the statefulset. The first pod would deploy successfully but the other
two would hang in a pending state. The reason for this is the `Volume
Binding Mode` of `immediate` needed to be configured to
`WaitForFirstConsumer`. This options changes the order in which the PV
and pod is created. With `WaitForFirstConsumer` the binding and
provisioning of a PersistentVolume is delayed until the pod using the
PersistentVolumeClaim  is created. Without this the Pod's scheduling
rules are not respected and the pod is unable to be both scheduled to the
node with the correct affinity and bind to the its PersistentVolume.
Therefore hanging in a pending state in perpetuity.

https://kubernetes.io/docs/concepts/storage/storage-classes/#volume-binding-mode