check & update Kubernetes and Docker tutorial
The Docker tutorial was updated to include more instructions on
installing Docker on Linux and to include the fact that Kubernetes moved
away from Docker.

In the Kubernetes training the api objects are adjusted to current
versions (k8s objects in the pg training too). Additionally instructions on how to execute the tutorial on
minikube are added, while a9s Paas references are removed.